+++ packages/eslint-plugin/src/rules/check-sql.rule.ts
@@ -1,10 +1,10 @@
 import { GenerateResult } from "@ts-safeql/generate";
-import { PostgresError } from "@ts-safeql/shared";
-import { ESLintUtils, TSESLint, TSESTree } from "@typescript-eslint/utils";
+import { objectKeysNonEmpty, PostgresError, defaultTypeMapping } from "@ts-safeql/shared";
+import { ESLintUtils, ParserServices, TSESLint, TSESTree } from "@typescript-eslint/utils";
 import pgParser from "libpg-query";
-import * as recast from "recast";
 import { createSyncFn } from "synckit";
 import { match } from "ts-pattern";
+import ts from "typescript";
 import z from "zod";
 import zodToJsonSchema from "zod-to-json-schema";
 import { ESTreeUtils } from "../utils";
@@ -12,12 +12,14 @@ import { E, flow, J, pipe } from "../utils/fp-ts";
 import { memoize } from "../utils/memoize";
 import { locateNearestPackageJsonDir } from "../utils/node.utils";
 import { mapTemplateLiteralToQueryText } from "../utils/ts-pg.utils";
+import { tsTypeToText } from "../utils/ts.utils";
 import { getConfigFromFileWithContext } from "./check-sql.config";
 import {
   reportBaseError,
   reportDuplicateColumns,
   reportIncorrectTypeAnnotations,
   reportInvalidQueryError,
+  reportInvalidTypeAnnotations,
   reportMissingTypeAnnotations,
   reportPostgresError,
   withTransformType,
@@ -28,8 +30,11 @@ const messages = {
   typeInferenceFailed: "Type inference failed {{error}}",
   error: "{{error}}",
   invalidQuery: "Invalid Query: {{error}}",
-  missingTypeAnnotations: "Query is missing type annotation",
-  incorrectTypeAnnotations: "Query has incorrect type annotation",
+  missingTypeAnnotations: "Query is missing type annotation\n\tFix with: {{fix}}",
+  incorrectTypeAnnotations:
+    "Query has incorrect type annotation.\n\tExpected: {{expected}}`\n\tActual: {{actual}}",
+  invalidTypeAnnotations:
+    "Query has invalid type annotation (SafeQL does not support it. If you think it should, please open an issue)",
 };
 export type RuleMessage = keyof typeof messages;
 
@@ -50,6 +55,16 @@ const baseSchema = z.object({
    * Whether or not keep the connection alive. Change it only if you know what you're doing.
    */
   keepAlive: z.boolean().optional(),
+
+  /**
+   * Override defaults
+   */
+  overrides: z
+    .object({
+      types: z.record(z.enum(objectKeysNonEmpty(defaultTypeMapping)), z.string()),
+    })
+    .partial()
+    .optional(),
 });
 
 const identifyByNameAndOperators = z.object({
@@ -227,10 +242,10 @@ function reportCheck(params: {
   tag: TSESTree.TaggedTemplateExpression;
   connection: RuleOptionConnection;
   projectDir: string;
-  typeParameters: TSESTree.TSTypeParameterInstantiation | undefined;
+  typeParameter: TSESTree.TSTypeParameterInstantiation | undefined;
   baseNode: TSESTree.BaseNode;
 }) {
-  const { context, tag, connection, projectDir, typeParameters, baseNode } = params;
+  const { context, tag, connection, projectDir, typeParameter, baseNode } = params;
 
   return pipe(
     E.Do,
@@ -240,7 +255,9 @@ function reportCheck(params: {
       mapTemplateLiteralToQueryText(tag.quasi, parser, checker)
     ),
     E.bindW("pgParsed", ({ query }) => pgParseQueryE(query)),
-    E.chainW(({ query, pgParsed }) => generateSyncE({ query, pgParsed, connection, projectDir })),
+    E.bindW("result", ({ query, pgParsed }) =>
+      generateSyncE({ query, pgParsed, connection, projectDir })
+    ),
     E.fold(
       (error) => {
         return match(error)
@@ -264,8 +281,8 @@ function reportCheck(params: {
           )
           .exhaustive();
       },
-      (result) => {
-        const isMissingTypeAnnotations = typeParameters === undefined;
+      ({ result, checker, parser }) => {
+        const isMissingTypeAnnotations = typeParameter === undefined;
         const resultWithTransformed = withTransformType(result, connection.transform);
 
         if (isMissingTypeAnnotations) {
@@ -277,11 +294,27 @@ function reportCheck(params: {
           });
         }
 
-        if (isIncorrectTypeAnnotations(resultWithTransformed, typeParameters)) {
+        const typeAnnotationState = getTypeAnnotationState({
+          result: resultWithTransformed,
+          typeParameter: typeParameter,
+          checker: checker,
+          parser: parser,
+        });
+
+        if (typeAnnotationState === "INVALID" || typeAnnotationState === "UNKNOWN") {
+          return reportInvalidTypeAnnotations({
+            context: context,
+            typeParameter: typeParameter,
+          });
+        }
+
+        if (!typeAnnotationState.isEqual) {
           return reportIncorrectTypeAnnotations({
             context,
             result: resultWithTransformed,
-            typeParameters: typeParameters,
+            typeParameter: typeParameter,
+            expected: typeAnnotationState.current,
+            actual: typeAnnotationState.generated,
           });
         }
       }
@@ -310,7 +343,7 @@ function checkConnectionByTagExpression(params: {
       connection,
       projectDir,
       baseNode: tag.tag,
-      typeParameters: tag.typeParameters,
+      typeParameter: tag.typeParameters,
     });
   }
 }
@@ -338,29 +371,52 @@ function checkConnectionByCallExpression(params: {
       connection,
       projectDir,
       baseNode: tag.parent.callee,
-      typeParameters: tag.parent.typeParameters,
+      typeParameter: tag.parent.typeParameters,
     });
   }
 }
 
-function isIncorrectTypeAnnotations(
-  { result }: GenerateResult,
-  sqlOperatorType: TSESTree.TSTypeParameterInstantiation
-) {
-  const currentType = recast.print(sqlOperatorType).code;
+function getTypeAnnotationState(params: {
+  result: GenerateResult;
+  typeParameter: TSESTree.TSTypeParameterInstantiation;
+  parser: ParserServices;
+  checker: ts.TypeChecker;
+}) {
+  const {
+    result: { result },
+    typeParameter,
+    parser,
+    checker,
+  } = params;
 
-  return !areTypesEqual(currentType, result);
+  if (typeParameter.params.length !== 1) {
+    return "INVALID" as const;
+  }
+
+  const typeNode = typeParameter.params[0];
+  const current = tsTypeToText({ checker, parser, typeNode });
+
+  if (current === "UNKNOWN") {
+    return "UNKNOWN" as const;
+  }
+
+  return getTypesEquality(current, result);
 }
 
 // TODO this should be improved.
-function areTypesEqual(current: string, generated: string | null) {
-  if (generated === null) {
-    return false;
+function getTypesEquality(current: string | null, generated: string | null) {
+  if (current === null && generated === null) {
+    return { isEqual: true, current, generated };
+  }
+
+  if (current === null || generated === null) {
+    return { isEqual: false, current, generated };
   }
 
   const omitRegex = /[\n ;]/g;
+  const isEqual = current.replace(omitRegex, "") === generated.replace(omitRegex, "");
 
-  return current.replace(omitRegex, "") === `<${generated}>`.replace(omitRegex, "");
+  return { isEqual, current, generated };
 }
 
 const createRule = ESLintUtils.RuleCreator(() => `https://github.com/ts-safeql/safeql`)<