+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ts-safeql/test-utils
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [69b874e]
+  - @ts-safeql/shared@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes